new update 20-01-2022

+++ MongoRestAPI/src/main/java/com/example/MongoRestAPI/EmployeeController.java
@@ -12,8 +12,12 @@ public class EmployeeController {
     @Autowired
     private EmployeeRepo repo;
 
+    @Autowired
+    private SequenceGeneratorService sequenceGeneratorService;
+
     @PostMapping("/addEmp")
     public String saveEmp(@RequestBody Employee emp){
+        emp.setEid(sequenceGeneratorService.generateSequence(Employee.SEQUENCE_NAME));
         repo.save(emp);
         return "Added Employee with id: " + emp.getEid();
     }
@@ -24,14 +28,24 @@ public class EmployeeController {
     }
 
     @GetMapping("/getEmp/{eid}")
-    public Optional<Employee> getEmployee(@PathVariable int eid){
-        return repo.findById(eid);
+    public String getEmployee(@PathVariable int eid){
+        if(repo.existsById(eid)){
+            return repo.findById(eid).toString();
+        }
+        else{
+            return "Given ID is not presented in Database Records!!!";
+        }
     }
 
     @DeleteMapping("/deleteEmp/{eid}")
     public String deleteEmp(@PathVariable int eid){
-        repo.deleteById(eid);
-        return "Record deleted for Employee with id: " + eid;
+        if(repo.existsById(eid)) {
+            repo.deleteById(eid);
+            return "Record deleted for Employee with id: " + eid;
+        }
+        else{
+            return "Record for this Id is already deleted or does not exists!!";
+        }
     }
 
     @PutMapping(path = "/updateEmp", consumes = { "application/json" })
